orchestration(ingame): Move renderer IPC into ingame API
